fix formula types and add tarantula (#243)
* fix formula types

* add tarantula

* add test for tarantula

* fix tarantula formula

* add math70 test for tarantula

+++ src/main/java/fr/spoonlabs/flacoco/localization/spectrum/formulas/OchiaiFormula.java
@@ -5,8 +5,8 @@ public class OchiaiFormula implements Formula {
 	public OchiaiFormula() {
 	}
 
-	public double compute(int nPassingNotExecuting, int nFailingNotExecuting, int nPassingExecuting,
-			int nFailingExecuting) {
+	public double compute(double nPassingNotExecuting, double nFailingNotExecuting, double nPassingExecuting,
+			double nFailingExecuting) {
 
 		if ((nFailingExecuting + nPassingExecuting == 0) || (nFailingExecuting + nFailingNotExecuting == 0)) {
 			return 0;
